+++ jina/__init__.py
@@ -3,7 +3,7 @@ __license__ = "Apache-2.0"
 
 # do not change this line manually
 # this is managed by git tag and updated on every release
-__version__ = '0.4.8'
+__version__ = '0.4.10'
 
 # do not change this line manually
 # this is managed by proto/build-proto.sh and updated on every execution
@@ -59,6 +59,7 @@ __jina_env__ = ('JINA_ARRAY_QUANT',
                 'JINA_TEST_CONTAINER',
                 'JINA_TEST_GPU',
                 'JINA_TEST_PRETRAINED',
+                'JINA_TEST_FAISS',
                 'JINA_VCS_VERSION',
                 'JINA_VERSION',
                 'JINA_WARN_UNNAMED',